+++ test/src/main/java/org/vertexium/test/GraphTestBase.java
@@ -190,11 +190,11 @@ public abstract class GraphTestBase {
     @Test
     public void testAddStreamingPropertyValue() throws IOException, InterruptedException {
         String expectedLargeValue = IOUtils.toString(new LargeStringInputStream(LARGE_PROPERTY_VALUE_SIZE));
-        PropertyValue propSmall = StreamingPropertyValue.create(new ByteArrayInputStream("value1".getBytes()), String.class, 6);
+        PropertyValue propSmall = StreamingPropertyValue.create(new ByteArrayInputStream("value1".getBytes()), String.class, 6L);
         PropertyValue propLarge = StreamingPropertyValue.create(
                 new ByteArrayInputStream(expectedLargeValue.getBytes()),
                 String.class,
-                expectedLargeValue.length()
+                null
         );
         String largePropertyName = "propLarge/\\*!@#$%^&*()[]{}|";
         Vertex v1 = graph.prepareVertex("v1", VISIBILITY_A)
@@ -279,7 +279,7 @@ public abstract class GraphTestBase {
         PropertyValue propLarge = StreamingPropertyValue.create(
                 new ByteArrayInputStream(expectedValue.getBytes()),
                 String.class,
-                expectedValue.length()
+                (long) expectedValue.length()
         );
         graph.prepareVertex("v1", VISIBILITY_A)
                 .addPropertyValue("key1", "largeProp", propLarge, metadata, timestamp, VISIBILITY_A)
@@ -295,7 +295,7 @@ public abstract class GraphTestBase {
         propLarge = StreamingPropertyValue.create(
                 new ByteArrayInputStream(expectedValue.getBytes()),
                 String.class,
-                expectedValue.length()
+                (long) expectedValue.length()
         );
         graph.prepareVertex("v1", VISIBILITY_A)
                 .addPropertyValue("key1", "largeProp", propLarge, metadata, timestamp + 1, VISIBILITY_A)
@@ -4484,8 +4484,11 @@ public abstract class GraphTestBase {
 
         Vertex v1 = graph.getVertex("v1", AUTHORIZATIONS_A_AND_B);
         StreamingPropertyValue spvA = (StreamingPropertyValue) v1.getPropertyValue("a");
+        assertEquals(12L, (long) spvA.getLength());
         StreamingPropertyValue spvB = (StreamingPropertyValue) v1.getPropertyValue("b");
+        assertEquals(12L, (long) spvA.getLength());
         StreamingPropertyValue spvC = (StreamingPropertyValue) v1.getPropertyValue("c");
+        assertEquals(12L, (long) spvA.getLength());
         ArrayList<StreamingPropertyValue> spvs = Lists.newArrayList(spvA, spvB, spvC);
         List<InputStream> streams = graph.getStreamingPropertyValueInputStreams(spvs);
         assertEquals("Test Value A", IOUtils.toString(streams.get(0)));